Logic lifted from CreateProfile to hooks/useCreateProfile
