+++ src/pages/CreateProfile.jsx
@@ -1,8 +1,4 @@
-import { useNavigate } from 'react-router-dom';
-import { useState } from 'react';
-
-// Data
-import { supabase } from '../lib/supabaseClient';
+import useCreateProfile from '../hooks/useCreateProfile';
 
 // Components
 import { ButtonRow } from "../components/buttons/ButtonRow";
@@ -21,63 +17,24 @@ import FishIcon from "../components/icons/Fish";
 import FrogIcon from "../components/icons/Frog";
 import TurtleIcon from "../components/icons/Turtle";
 import { ErrorModal } from '../components/ErrorModal';
+    
 
 export default function CreateProfile() {
 
-    const navigate = useNavigate();
-    const [errorMessage, setErrorMessage] = useState("");
-
-    // Fetch profile
-    const savedProfileText = localStorage.getItem('userProfile');
-    const existingProfile = savedProfileText ? JSON.parse(savedProfileText) : null;
-
-    // Form data
-    const [name, setName] = useState(() => {
-        const saved = localStorage.getItem('userProfile');
-        return saved ? JSON.parse(saved).name : "";
-    });
-    const [role, setRole] = useState(() => {
-        const saved = localStorage.getItem('userProfile');
-        return saved ? JSON.parse(saved).role : "";
-    });
-    const [avatar, setAvatar] = useState(() => {
-        const saved = localStorage.getItem('userProfile');
-        return saved ? JSON.parse(saved).avatar : "";
-    });
-    const [loading, setLoading] = useState(false);
-
-    
-    // Submit function
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-        setLoading(true);
-
-        const profileData = { 
-            name: existingProfile ? existingProfile.name : name, 
-            role, 
-            avatar };
-
-        try {
-            const { data, error } = await supabase
-                .from('users')
-                .insert([profileData]);
-
-            if(error) throw error;
-
-            // Save to local storage
-            localStorage.setItem('userProfile', JSON.stringify(profileData));
-
-            console.log("Profil sparad!", data);
-            navigate('/finished-profile')
-
-        } catch (error) {
-            console.error("Fel vid uppladdning: ", error.message);
-            setErrorMessage(error.message);
-
-        } finally {
-            setLoading(false)
-        }
-    }
+    const {
+        name,
+        setName,
+        role,
+        setRole,
+        avatar,
+        setAvatar,
+        loading,
+        errorMessage,
+        setErrorMessage,
+        existingProfile,
+        handleSubmit,
+        navigate,
+    } = useCreateProfile();
 
     return (
         <>
@@ -226,7 +183,7 @@ export default function CreateProfile() {
                     TILLBAKA
                 </SmallLightButton>
                 <SmallButton type="submit" disabled={loading}>
-                    {loading ? 'SPARAR...' : 'KLAR'}
+                    {loading ? 'SPARAR...' : 'SPARA'}
                     <img src="forwardArrow.svg" alt="forward" />
                 </SmallButton>
             </ButtonRow>
